+++ cmd/registry-storagedriver-filesystem/main.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/Sirupsen/logrus"
 
-	"github.com/docker/distribution/storagedriver/filesystem"
-	"github.com/docker/distribution/storagedriver/ipc"
+	"github.com/docker/distribution/registry/storage/driver/filesystem"
+	"github.com/docker/distribution/registry/storage/driver/ipc"
 )
 
 // An out-of-process filesystem driver, intended to be run by ipc.NewDriverClient
