+++ registry/storage/driver/azure/azure.go
@@ -11,9 +11,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/docker/distribution/storagedriver"
-	"github.com/docker/distribution/storagedriver/base"
-	"github.com/docker/distribution/storagedriver/factory"
+	storagedriver "github.com/docker/distribution/registry/storage/driver"
+	"github.com/docker/distribution/registry/storage/driver/base"
+	"github.com/docker/distribution/registry/storage/driver/factory"
 
 	azure "github.com/MSOpenTech/azure-sdk-for-go/clients/storage"
 )
