+++ registry/storage/driver/azure/azure_test.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/docker/distribution/storagedriver"
-	"github.com/docker/distribution/storagedriver/testsuites"
+	storagedriver "github.com/docker/distribution/registry/storage/driver"
+	"github.com/docker/distribution/registry/storage/driver/testsuites"
 	. "gopkg.in/check.v1"
 )
