+++ registry/storage/driver/azure/blockblob.go
