+++ registry/storage/driver/azure/randomwriter.go
