+++ registry/storage/driver/azure/randomwriter_test.go
