+++ registry/storage/driver/azure/zerofillwriter.go
