+++ registry/storage/driver/azure/zerofillwriter_test.go
