+++ registry/storage/driver/filesystem/driver_test.go
@@ -5,8 +5,8 @@ import (
 	"os"
 	"testing"
 
-	"github.com/docker/distribution/storagedriver"
-	"github.com/docker/distribution/storagedriver/testsuites"
+	storagedriver "github.com/docker/distribution/registry/storage/driver"
+	"github.com/docker/distribution/registry/storage/driver/testsuites"
 	. "gopkg.in/check.v1"
 )
