+++ registry/storage/driver/inmemory/driver_test.go
@@ -3,8 +3,8 @@ package inmemory
 import (
 	"testing"
 
-	"github.com/docker/distribution/storagedriver"
-	"github.com/docker/distribution/storagedriver/testsuites"
+	storagedriver "github.com/docker/distribution/registry/storage/driver"
+	"github.com/docker/distribution/registry/storage/driver/testsuites"
 
 	"gopkg.in/check.v1"
 )
