+++ registry/storage/driver/ipc/client.go
@@ -13,7 +13,7 @@ import (
 	"os/exec"
 	"syscall"
 
-	"github.com/docker/distribution/storagedriver"
+	storagedriver "github.com/docker/distribution/registry/storage/driver"
 	"github.com/docker/libchan"
 	"github.com/docker/libchan/spdy"
 )
