+++ registry/storage/driver/ipc/server.go
@@ -10,7 +10,7 @@ import (
 	"os"
 	"reflect"
 
-	"github.com/docker/distribution/storagedriver"
+	storagedriver "github.com/docker/distribution/registry/storage/driver"
 	"github.com/docker/libchan"
 	"github.com/docker/libchan/spdy"
 )
