+++ registry/storage/driver/s3/s3.go
@@ -26,9 +26,9 @@ import (
 
 	"github.com/AdRoll/goamz/aws"
 	"github.com/AdRoll/goamz/s3"
-	"github.com/docker/distribution/storagedriver"
-	"github.com/docker/distribution/storagedriver/base"
-	"github.com/docker/distribution/storagedriver/factory"
+	storagedriver "github.com/docker/distribution/registry/storage/driver"
+	"github.com/docker/distribution/registry/storage/driver/base"
+	"github.com/docker/distribution/registry/storage/driver/factory"
 )
 
 const driverName = "s3"
