+++ registry/storage/driver/storagedriver.go
@@ -1,4 +1,4 @@
-package storagedriver
+package driver
 
 import (
 	"errors"
